requirements.txt
APScheduler==3.10.4
bidict==0.22.1
bottle==0.12.25
dnspython==2.4.2
frozenlist==1.4.0
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==21.2.0
jsonpickle==3.0.2
pipdeptree==2.13.0
pyhumps==3.8.0
python-dotenv==1.0.0
python-engineio==4.6.1
python-socketio==5.8.0
pytz==2023.3
six==1.16.0
tzlocal==5.0.1
zope.event==5.0
zope.interface==6.0
